Conditionalize mounting
It looks like, starting in either 4.17 or 4.18, the kernel
gets *really* unhappy about the second bind mount request.  In the
past it was a no-op; now it seems to perform a second mount.  While
only slightly irritating on its face, this also has the effect of
failing all subsequent PTY requests.  So... you can't open any more
terminals.  (su works, sudo doesn't.)

Make a low-effort guess as to whether we've mounted already, and if we
have, don't try again.  Since it's pretty obvious if we don't set up
everything correctly, work harder at writing this commit message than
the actual checks.

Also, put in tons of f-strings, because this is python3.